Move page parsing logic to helper
