+++ ynr/apps/sopn_parsing/management/commands/sopn_parsing_extract_page_numbers.py
@@ -1,7 +1,10 @@
 from django.core.management.base import BaseCommand
 
 from official_documents.models import OfficialDocument
-from sopn_parsing.helpers.pdf_helpers import NoTextInDocumentError, SOPNDocument
+from sopn_parsing.helpers.extract_pages import (
+    save_page_numbers_for_single_document,
+)
+from sopn_parsing.helpers.text_helpers import NoTextInDocumentError
 
 
 class Command(BaseCommand):
@@ -40,48 +43,10 @@ class Command(BaseCommand):
         qs = OfficialDocument.objects.filter(**filter_kwargs)
         seen_sources = set()
         for document in qs:
-
             if document.source_url in seen_sources:
                 continue
-
-            other_doc_models = (
-                OfficialDocument.objects.filter(source_url=document.source_url)
-                .exclude(pk=document.pk)
-                .select_related("post_election", "post_election__post")
-            )
-
             try:
-                for other_doc, pages in self.parse_single_document(
-                    document, other_doc_models
-                ):
-                    other_doc.relevant_pages = pages
-                    other_doc.save()
-            except ValueError as e:
+                save_page_numbers_for_single_document(document)
+            except (ValueError, NoTextInDocumentError) as e:
                 self.stderr.write(e)
             seen_sources.add(document.source_url)
-
-    def parse_single_document(self, document, other_doc_models):
-        self.top_strings = []
-
-        filename = document.uploaded_file.path
-        if not filename:
-            return
-
-        if not other_doc_models.exists():
-            yield document, "all"
-            return
-        try:
-            sopn = SOPNDocument(filename)
-        except (NoTextInDocumentError):
-            self.stdout.write("No text in {}, skipping".format(filename))
-            return
-        for other_doc in other_doc_models:
-            pages = sopn.get_pages_by_ward_name(
-                other_doc.post_election.post.label
-            )
-            if not pages:
-                raise ValueError(
-                    "None of the ballots fund in file {}".format(filename)
-                )
-            page_numbers = ",".join(str(p.page_number) for p in pages)
-            yield other_doc, page_numbers
